Asset manager service implemented in service

+++ src/main/java/snippet_manager/snippet/services/CodeSnippetService.java
@@ -4,14 +4,19 @@ import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.PermissionDeniedDataAccessException;
 import org.springframework.http.ResponseEntity;
+import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpServerErrorException;
+import org.springframework.web.multipart.MultipartFile;
 import snippet_manager.snippet.model.dtos.CodeSnippetDTO;
 import snippet_manager.snippet.model.entities.CodeSnippet;
+import snippet_manager.snippet.webservice.asset.AssetManager;
 import snippet_manager.snippet.webservice.permission.PermissionManager;
 import snippet_manager.snippet.repositories.CodeSnippetRepository;
 import snippet_manager.snippet.webservice.printscript.PrintscriptManager;
 
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -27,19 +32,29 @@ public class CodeSnippetService {
   @Autowired
   private PrintscriptManager printscriptManager;
 
-  public String createSnippet(CodeSnippetDTO snippet, Long userId) {
+  @Autowired
+  private AssetManager assetManager;
+
+  public String createSnippet(CodeSnippetDTO snippet, Long userId) throws IOException {
     compileSnippet(snippet);
     CodeSnippet codeSnippet = new CodeSnippet();
-    codeSnippet.setTitle(snippet.getTitle());
-    codeSnippet.setContent(snippet.getContentInString());
+    codeSnippet.setAssetId(snippet.getAssetId());
     codeSnippet.setLanguage(snippet.getLanguageInEnum());
     codeSnippet.setVersion(snippet.getVersion());
     codeSnippetRepository.save(codeSnippet);
 
     ResponseEntity<String> permissionResponse = createNewPermission(userId, codeSnippet.getId());
     if (permissionResponse.getStatusCode().isError()) {
+      codeSnippetRepository.deleteById(codeSnippet.getId());
       throw new HttpServerErrorException(permissionResponse.getStatusCode());
     }
+
+    ResponseEntity assetResponse = assetManager.createAsset("snippets", snippet.getAssetId(), snippet.getContent());
+    if (assetResponse.getStatusCode().isError()) {
+      codeSnippetRepository.deleteById(codeSnippet.getId());
+      throw new HttpServerErrorException(assetResponse.getStatusCode());
+    }
+
     return "Snippet created successfully";
   }
 
@@ -50,7 +65,7 @@ public class CodeSnippetService {
     }
   }
 
-  public CodeSnippetDTO getSnippet(UUID snippetId, Long userId) {
+  public CodeSnippetDTO getSnippet(UUID snippetId, Long userId) throws IOException {
     boolean canAccess = canReadSnippet(userId, snippetId);
     if (!canAccess) {
       throw new PermissionDeniedDataAccessException("You don't have permission to access this snippet", new Exception("You don't have permission to access this snippet"));
@@ -58,34 +73,48 @@ public class CodeSnippetService {
     CodeSnippet codeSnippet = codeSnippetRepository.findById(snippetId)
             .orElseThrow(() -> new EntityNotFoundException("Snippet not found with id " + snippetId));
 
-    return CodeSnippetDTO.builder()
-        .title(codeSnippet.getTitle())
-        .language(codeSnippet.getLanguage().name())
-        .content(codeSnippet.getContentInMultipartFile())
-        .version(codeSnippet.getVersion())
-        .build();
+    InputStream assetResponse = assetManager.getAsset("snippets", codeSnippet.getAssetId());
+    MultipartFile asset = toMultipartFile(assetResponse, codeSnippet.getAssetId());
+    return convertToDTO(codeSnippet, asset);
+  }
+
+  private MultipartFile toMultipartFile(InputStream assetResponse, String fileName) throws IOException {
+    byte[] bytes = assetResponse.readAllBytes();
+    return new MockMultipartFile(fileName, fileName, "text/plain" , bytes);
   }
 
   public List<CodeSnippetDTO> getAllSnippets(Long userId) {
     List<CodeSnippet> codeSnippets = codeSnippetRepository.findAll();
-    for (CodeSnippet codeSnippet : codeSnippets) {
-      boolean canAccess = canReadSnippet(userId, codeSnippet.getId());
-      if (!canAccess) {
-        codeSnippets.remove(codeSnippet);
-      }
-    }
-    return convertToDTO(codeSnippets);
+
+    List<CodeSnippetDTO> codeSnippetDTOS = codeSnippets.stream()
+            .filter(codeSnippet -> canReadSnippet(userId, codeSnippet.getId()))
+            .map(codeSnippet -> {
+              try {
+                InputStream assetResponse = assetManager.getAsset("snippets", codeSnippet.getAssetId());
+                MultipartFile asset = toMultipartFile(assetResponse, codeSnippet.getAssetId());
+
+                return convertToDTO(codeSnippet, asset);
+              } catch (IOException e) {
+                throw new RuntimeException("Error reading the file content: " + codeSnippet.getId(), e);
+              }
+            })
+            .collect(Collectors.toList());
+
+    return codeSnippetDTOS;
   }
 
-  public String updateSnippet(UUID snippetId, Long userId, CodeSnippetDTO codeSnippet) {
+  public String updateSnippet(UUID snippetId, Long userId, CodeSnippetDTO codeSnippet) throws IOException {
     boolean canAccess = canWriteSnippet(userId, snippetId);
     if (!canAccess) {
       throw new PermissionDeniedDataAccessException("You don't have permission to access this snippet", new Exception("You don't have permission to access this snippet"));
     }
     CodeSnippet existingCodeSnippet = codeSnippetRepository.findById(snippetId)
             .orElseThrow(() -> new EntityNotFoundException("Snippet not found with id " + snippetId));
-    existingCodeSnippet.setTitle(codeSnippet.getTitle());
-    existingCodeSnippet.setContent(codeSnippet.getContentInString());
+
+    ResponseEntity<?> assetResponse = assetManager.createAsset("snippets", existingCodeSnippet.getAssetId(), codeSnippet.getContent());
+    if (assetResponse.getStatusCode().isError()) {
+      throw new HttpServerErrorException(assetResponse.getStatusCode());
+    }
     existingCodeSnippet.setLanguage(codeSnippet.getLanguageInEnum());
     existingCodeSnippet.setVersion(codeSnippet.getVersion());
     codeSnippetRepository.save(existingCodeSnippet);
@@ -100,21 +129,21 @@ public class CodeSnippetService {
     codeSnippetRepository.findById(snippetId)
             .orElseThrow(() -> new EntityNotFoundException("Snippet not found with id " + snippetId));
 
+    ResponseEntity<?> assetResponse = assetManager.deleteAsset("snippets", snippetId.toString());
+    if (assetResponse.getStatusCode().isError()) {
+      throw new HttpServerErrorException(assetResponse.getStatusCode());
+    }
     codeSnippetRepository.deleteById(snippetId);
     return "Snippet deleted successfully";
   }
 
   //Internal methods
-
-  private List<CodeSnippetDTO> convertToDTO(List<CodeSnippet> codeSnippets) {
-    return codeSnippets.stream()
-            .map(codeSnippet -> CodeSnippetDTO.builder()
-                    .title(codeSnippet.getTitle())
-                    .language(codeSnippet.getLanguage().name())
-                    .content(codeSnippet.getContentInMultipartFile())
-                    .version(codeSnippet.getVersion())
-                    .build())
-            .collect(Collectors.toList());
+  private CodeSnippetDTO convertToDTO(CodeSnippet codeSnippets, MultipartFile content) {
+    return CodeSnippetDTO.builder()
+            .language(codeSnippets.getLanguage().name())
+            .version(codeSnippets.getVersion())
+            .content(content)
+            .build();
   }
 
   private ResponseEntity<String> createNewPermission(Long userId, UUID snippetId) {
